use unsigned long for start ms
i changed the duration to 5000 to test but it seems to end up with the rumble never finishing.

thought it might be due to this, due to arduino documentation mentioning "logic errors may occur if a programmer tries to do arithmetic with smaller data types such as int" but sadly doesn't seem to fix

https://docs.arduino.cc/language-reference/en/functions/time/millis/

+++ Arduino/imu/bluetooth_imu/bluetooth_imu.ino
@@ -20,8 +20,8 @@ const int motor2SPDPin = A1;  // Motor 2 Speed PWM pin of dual motor driver conn
 const int motor2DIRPin = A0;  // Motor 2 Direction pin of dual motor driver connected to digital pin 26
 
 bool rumbleOn = false;
-int rumbleStartMs;
-int rumbleDuration;
+unsigned long rumbleStartMs;
+unsigned long rumbleDuration;
 
 Adafruit_BNO08x bno08x(BNO08X_RESET);
 sh2_SensorValue_t sensorValue;
@@ -113,6 +113,7 @@ void startRumble(int duration) {
 
   rumbleOn = true;
   rumbleStartMs = millis();
+  //NOTE: assume duration is unsigned
   rumbleDuration = duration;
   Serial1.println("Rumble activated.");
